[permissions] Add script to visualise permission relationships. Fixes JB#52333
This script allows to visualise permission relationships. It mostly
deals with includes of other permissions. This is useful to tell if
there are problematic implicit dependencies.

It also shows if permission can be directly included or if it's so
called internal permission. That heuristic is based on missing
translation category which is mostly correct indicator (there may be
false negatives).

Additionally it shows if permission defines privileged data directory.

Graphviz is required to use this tool.

Signed-off-by: Tomi Leppänen <tomi.leppanen@jolla.com>