Fix node coloring in dependencies.py
I noticed that when nodes are iterated in very specific order,
non-privileged nodes may be colored lightgray which should be marking
nodes that are privileged. This happened when privileged node was
followed by non-privileged child nodes.

It's not clear if this is a problem with graphviz library or mistake in
usage of the API. In any case this fixes the issue by moving setting
style and color attributes to node creation.

Signed-off-by: Tomi Leppänen <tomi.leppanen@jolla.com>