[permissions] Move pin query to its own permission. JB#53821 OMP#JOLLA-53
Move pin query dialog opening to its own permission but include that
from Phone and Messages. Also restrict this to pin query only as that
service name exposes some other interfaces as well.

Add a comment that this D-Bus API requires privileged access to
function. This permission is internal without any descriptions. As this
is only used through Phone and Messages those are not needed.

Signed-off-by: Tomi Leppänen <tomi.leppanen@jolla.com>