[permissions] Restrict audio permissions, update Microphone perms.  OMP#JOLLA-50 JB#53817
Use "nosound" to disable access to any sound device unless Audio
permission is used.

Update to Microphone permission to include Audio permission. There are
no means to separate Audio permission to playback and record, so this
includes Audio permission by default in order to allow recording with
this permission. As a downside audio can be also played. This is,
therefore, same as including Audio permission but meant to be used as a
future permission when the issue of separating permissions to playback
and record can be done in Pulseaudio.

Expand and update the Audio and Microphone descriptions about Pulsaudio.