[permissions] Abstract Telepathy access for messages. OMP#JOLLA-88 JB#54034
Abstract the Telepathy access for messages by using an unique name that
is formed of organization and application name in the D-Bus name
ownership. These values are supported via --template in sailjail that
will get the unique name name passed and processed by firejail in
addition to organization and application names.

Move the D-Bus name ownership from jolla-messages app specific profile
to the Messages.permission for all messaging apps to use.

Signed-off-by: Jussi Laakkonen <jussi.laakkonen@jolla.com>