[permissions] Update README to clarify which apps are sandboxed. Fixes JB#58455
The README suggested there was a plan to sandbox all apps by the
Sailfish OS 4.4.0 release. In practice, this meant that all apps without
permissions specified would be given a default profile from 4.4.0
onwards.

Since this is now the case, this change updates the README to reflect
the current situation.